2015-02-24 AEP Edits negative A and Z
MSYL = 0.3 wasn't working, because A and Z parameters were both negative
(it is a VERY weird density-dependence function indeed).
AEP log-transformed A and Z so it was impossible to solve for
negative A and Z. The previous error is now corrected for by AEP and this
will embark the first round of testing to make sure the changes worked
and are leading towards better results.

Modified MANTSALT.FOR to allow for negative A and Z