New MANTST14 w/ comments from AEP
There are some additional changes, such as line 1469 and 1919.
The changes here may need to be fixed, I am not sure yet. Prior the
line on 1469 was commented out and called MSYPAR.
Also line 1919 was IYR.GT.0 not IYR.GE.0

I will run and see if these need to be changed.