Changes to MSYPARZZ by Andre
Changes were done so that MSYPARZZ matches results from
Norway that were were compiled by Ragnar. This work
was done in 2007 but went to us in March of 2015. The RES
outputs now match within the limits of rounding error.
Need to go back and comment out the MSYPARZZ and to leave the
MSYPAR2 as the used routine.