Pass starting values for A and Z.
COPY.DAT now contains starting values for A and Z,
which are passed to MSYPAR2 and subsequently used
in the loop to find the best values for A and Z.
In doing this I noticed an error, that has yet to be
fixed, in the dat file. The values of KYEAR and MSYRYR
are read in, but the placement of the conditional
statement that reads them in is not in the correct
spot in MANTST14.FOR. This needs to be fixed, but right
now is not an issue b/c the conditional statement does
not lead to reading in the values.