MANSRESV9.for added output statistics needed to compare CLAs
All information was there, AEP just needed to find it and output it.
The program now outputs the original file RESOUT.RRR and an additional
file RESOUT2.RRR where there are more, I think, summary statistics, and
the file is in a different structure. This may change how things are
read in and also how they are to be organized before they are plotted.